getting out IntervalOverlappingIterator from LocusWalker
